Do not pretend Matrix can have bigger indexes than int

+++ src/map/matrix.h
@@ -1,7 +1,6 @@
 #ifndef MATRIX_H
 #define MATRIX_H
 
-#include <cstddef>
 #include <cfloat>
 #include <QVector>
 #include <QDebug>
@@ -10,12 +9,12 @@ class Matrix
 {
 public:
 	Matrix() {_h = 0; _w = 0;}
-	Matrix(size_t h, size_t w);
+	Matrix(int h, int w);
 
-	size_t h() const {return _h;}
-	size_t w() const {return _w;}
-	double &m(size_t i, size_t j) {return _m[_w * i + j];}
-	double const &m(size_t i, size_t j) const {return _m.at(_w * i + j);}
+	int h() const {return _h;}
+	int w() const {return _w;}
+	double &m(int i, int j) {return _m[_w * i + j];}
+	double const &m(int i, int j) const {return _m.at(_w * i + j);}
 
 	bool isNull() const {return (_h == 0 || _w == 0);}
 
@@ -24,8 +23,8 @@ public:
 
 private:
 	QVector<double> _m;
-	size_t _h;
-	size_t _w;
+	int _h;
+	int _w;
 };
 
 #ifndef QT_NO_DEBUG
